+++ vscode/webviews/chat/Transcript.story.tsx
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from '@storybook/react'
 
 import { Transcript } from './Transcript'
-import { FIXTURE_TRANSCRIPT, FIXTURE_USER_ACCOUNT_INFO } from './fixtures'
+import { FIXTURE_TRANSCRIPT, FIXTURE_USER_ACCOUNT_INFO, transcriptFixture } from './fixtures'
 
 import { RateLimitError, errorToChatError, ps } from '@sourcegraph/cody-shared'
 import type { ComponentProps } from 'react'
@@ -84,44 +84,45 @@ export const WaitingForContext: StoryObj<typeof meta> = {
 
 export const WaitingForAssistantMessageWithContext: StoryObj<typeof meta> = {
     args: {
-        transcript: [
+        transcript: transcriptFixture([
             ...SIMPLE_TRANSCRIPT,
             {
                 speaker: 'human',
                 text: ps`What color is the sky?'`,
                 contextFiles: [{ type: 'file', uri: URI.file('/foo.js') }],
             },
-        ],
-        messageInProgress: { speaker: 'assistant' },
+        ]),
+        messageInProgress: { speaker: 'assistant', model: 'my-llm' },
     },
 }
 
 export const WaitingForAssistantMessageNoContext: StoryObj<typeof meta> = {
     args: {
-        transcript: [
+        transcript: transcriptFixture([
             ...SIMPLE_TRANSCRIPT,
             {
                 speaker: 'human',
                 text: ps`What color is the sky?'`,
                 contextFiles: [],
             },
-        ],
+        ]),
         messageInProgress: { speaker: 'assistant' },
     },
 }
 
 export const AssistantMessageInProgress: StoryObj<typeof meta> = {
     args: {
-        transcript: [
+        transcript: transcriptFixture([
             ...SIMPLE_TRANSCRIPT,
             {
                 speaker: 'human',
                 text: ps`What color is the sky?'`,
                 contextFiles: [{ type: 'file', uri: URI.file('/foo.js') }],
             },
-        ],
+        ]),
         messageInProgress: {
             speaker: 'assistant',
+            model: 'my-model',
             text: ps`The sky is `,
         },
     },
@@ -129,18 +130,18 @@ export const AssistantMessageInProgress: StoryObj<typeof meta> = {
 
 export const WithError: StoryObj<typeof meta> = {
     args: {
-        transcript: [
+        transcript: transcriptFixture([
             ...SIMPLE_TRANSCRIPT,
             { speaker: 'human', text: ps`What color is the sky?'`, contextFiles: [] },
             { speaker: 'assistant', error: errorToChatError(new Error('some error')) },
-        ],
+        ]),
         isTranscriptError: true,
     },
 }
 
 export const WithRateLimitError: StoryObj<typeof meta> = {
     args: {
-        transcript: [
+        transcript: transcriptFixture([
             ...SIMPLE_TRANSCRIPT,
             { speaker: 'human', text: ps`What color is the sky?'`, contextFiles: [] },
             {
@@ -149,7 +150,7 @@ export const WithRateLimitError: StoryObj<typeof meta> = {
                     new RateLimitError('chat messages and commands', 'rate limit error', true)
                 ),
             },
-        ],
+        ]),
         isTranscriptError: true,
     },
 }